Extract Branch and Line classes to separate file
While we're on the topic of cleaning up source file - get
the line and branch classes to their own respective separate
files.
